checkpoint: migrates implementation from "library/Schema" to "library/Sequence"
