+++ src/library/Schema/base64CharacterEncodedByteSequence.ts
@@ -1,25 +1,6 @@
 import Sequence from '@/library/Sequence';
 
-import {
-  string as Schema_string,
-  NEVER as Schema_NEVER,
-} from './core';
-
-/** An alternative to `Schema.base64()` that avoids using the deprecated `atob` conversion under the hood */
-const Schema_base64CharacterEncodedByteSequence = () => Schema_string().transform((someSubject, currentContext) => {
-  const outcomeOfParsingSubject = Sequence.Byte.Encoded.Base64.parsedFrom(someSubject);
-
-  if (
-    outcomeOfParsingSubject.isSuccess
-  ) return outcomeOfParsingSubject.unwrapped;
-
-  currentContext.addIssue({
-    code   : 'custom',
-    message: outcomeOfParsingSubject.cause.message,
-  });
-
-  return Schema_NEVER;
-});
+const Schema_base64CharacterEncodedByteSequence = () => Sequence.Byte.Encoded.Base64.Schema();
 
 export {
   Schema_base64CharacterEncodedByteSequence,
